+++ electron/project_controller.ts
@@ -1,46 +1,55 @@
 import { ipcMain } from 'electron';
 import * as db from './db.js';
+import type { Layout } from 'react-grid-layout';
 
 type response = {
-    success: boolean;
     data: any;
+    err: any;
 };
 // Returns all widgets for given project id
 // Layout is included if record is found
 export function InitProjectController() : void {
-    ipcMain.handle('get-project-widgets', async (event, args : any) : Promise<response> => {
+    ipcMain.handle('get-project-widgets', async (event, args : {projectId : number}) : Promise<response> => {
         try {
-            const widgets : db.widget[] = await db.GetWidgets(args.projectId);
-            return { success: true, data: widgets }
+            const widgetData = await db.GetWidgets(args.projectId);
+            return { data: widgetData, err: null }
         } catch (err) {
-            return { success: false, data: err }
+            return { data: null, err: err }
         }
     });
 }
 // Returns all projects for given username
-ipcMain.handle('get-user-projects', async (event, args : any) : Promise<response> => {
+ipcMain.handle('get-user-projects', async (event, args : {
+    username : string
+}) : Promise<response> => {
     try {
         const projects : db.project[] = await db.GetProjects(args.username);
-        return { success: true, data: projects };
+        return { data: projects, err: null };
     } catch (err : any) {
-        return { success: false, data: err };
+        return { data: null, err: err };
     }
 });
-
-ipcMain.handle('create-widget', async (event, args : any) : Promise<response> => {
+// Creates a widget
+ipcMain.handle('create-widget', async (event, args : {
+    projectId : number,
+    name : string,
+    layout : Layout
+}) : Promise<response> => {
     try {
         const id : number = await db.CreateWidget(args.projectId, args.name, args.layout);
-        return { success: true, data: id };
+        return { data: id, err: null };
     } catch (err : any) {
-        return { success: false, data: err };
+        return { data: null, err: err };
     }
 });
-
-ipcMain.handle('update-all-widget-layouts', async (event, args : any) : Promise<response> => {
+// Updates the position of the entire grid
+ipcMain.handle('update-all-widget-layouts', async (event, args : {
+    grid : Layout[]
+}) : Promise<response> => {
     try {
         db.UpdateAllWidgetLayouts(args.grid);
-        return { success: true, data: null };
+        return { data: null, err: null };
     } catch (err : any) {
-        return {success: false, data: err };
+        return { data: null, err: err };
     }
 });